feat(mini-mio): initial commit
This initial commit includes the first phase of the
work I was doing to recreate a similar / minimal
version of mio. This was just for learning purposes to
understand how to create ffi's to epoll and use epoll
to create an event queue, register interest for events
and handle events upon receiving notifications.

A delayserver is also required and was placed in
`rust-async-utils`.