feat(runtime): initial implementation of a simple runtime
This is a simplified runtime. The reactor (Poll) and executor
(scheduler) are being run on the same thread. It does however differ
from `a-coroutine` in that we do not continuously poll the futures to
see if they are ready. Rather, we use `mio::Poll` to register an event
source and have our runtime block on events received from the OS event
queue. This makes it more efficient than the `a-coroutine`
implementation, but still not a fully functional runtime.

The executor is also limited to polling 1 top-level future.

Overall, there is a tight integration between the reactor and the
executor as both must be aware of `mio::Poll`. This means we can't run
multiple different reactors that wake up the same executor for different
reasons.