feat(b-reactor-executor): bootstrap new binary
This builds upon work done in `a-runtime`. The main goal is to remove
the tight coupling between the reactor and executor via introducing
usage of a Waker.

This initial commit is simple bootstrapping the new binary using
`a-runtime` + minor restructuring of `runtime.rs` into a directory
module, with separate `executor.rs` and `reactor.rs` modules.