feat(b-reactor-executor)!: initial Executor implementation
Very basic Executor that allows multiple top-level futures (also known
as tasks) to be driven to completion. The Executor is really split into
two halves. State is stored in the thread local static `ExecutorCore`,
and `Executor` is just a unit struct with helper methods around this
`ExecutorCore`. If multiple executors are run, they are not aware of
each other or the tasks in other executors: hence, this is a not a work
stealing implementation, as tasks cannot be resumed / moved across
threads.

BREAKING CHANGES: still will not compile as Reactor implementation is
not complete.