feat(c-coroutines-problem): introduce self-reference error
An optimisation is made here to poll a future once and if it returns
ready, then we can complete early. However, this introduces a situation
whereby by polling the future we transition it to a state where one of
it's fields holds a reference to a reference to another field in the
future - i.e. a self reference is kept across yield points. When the
future is next `moved` into the `spawn` function, this invalidates the
references being held and we get a OS error when we next try to attempt
to dereference pointer within the future.

This is simply an initial example that leads into some exploration
around pinning.